+++ packages/typespec-python/src/lib.ts
@@ -10,7 +10,6 @@ export interface PythonEmitterOptions {
     "package-pprint-name"?: string;
     "head-as-boolean"?: boolean;
     "models-mode"?: string;
-    "unbranded"?: boolean;
     "tracing"?: boolean;
     "company-name"?: string;
     "debug"?: boolean;
@@ -30,7 +29,6 @@ const EmitterOptionsSchema: JSONSchemaType<PythonEmitterOptions> = {
         "package-pprint-name": { type: "string", nullable: true },
         "head-as-boolean": { type: "boolean", nullable: true },
         "models-mode": { type: "string", nullable: true },
-        "unbranded": { type: "boolean", nullable: true },
         "tracing": { type: "boolean", nullable: true },
         "company-name": { type: "string", nullable: true },
         "debug": { type: "boolean", nullable: true },